+++ package.json
@@ -19,6 +19,7 @@
     "react-router-dom": "^6.12.1",
     "react-scripts": "5.0.1",
     "react-simple-star-rating": "^5.1.7",
+    "recoil": "^0.7.7",
     "styled-components": "^5.3.10",
     "styled-reset": "^4.4.7",
     "use-debounce": "^9.0.4",
