+++ src/atoms/randomFoodAtom.js
@@ -1,0 +1,11 @@
+import { atom } from "recoil";
+
+export const randomFoodState = atom({
+  key: "randomFoodState",
+  default: "",
+});
+
+export const isAnimationActiveState = atom({
+  key: "isAnimationActiveState",
+  default: false,
+});
